Introduce TPMDevicePropertyDecoder
to make room to customize how to decode TPMDeviceProperty from subject
alternative name
