+++ webauthn4j-core/src/main/java/com/webauthn4j/validator/attestation/statement/tpm/TPMDevicePropertyDecoder.java
@@ -1,0 +1,8 @@
+package com.webauthn4j.validator.attestation.statement.tpm;
+
+import com.webauthn4j.validator.exception.BadAttestationStatementException;
+
+public interface TPMDevicePropertyDecoder {
+
+    TPMDeviceProperty decode(String subjectAlternativeName) throws BadAttestationStatementException;
+}
